algoritm/python_project/requirements.txt
Flask==3.0.3
numpy==2.1.1
pillow==10.4.0
